+++ tests/pools/abstract-pool.test.mjs
@@ -826,6 +826,7 @@ Deno.test({
           expect(workerNode.tasksQueue).toBeInstanceOf(PriorityQueue)
           expect(workerNode.tasksQueue.size).toBe(0)
           expect(workerNode.tasksQueue.maxSize).toBe(0)
+          expect(workerNode.tasksQueue.k).toBe(numberOfWorkers * 2)
         }
         await pool.destroy()
         pool = new DynamicThreadPool(
@@ -838,6 +839,7 @@ Deno.test({
           expect(workerNode.tasksQueue).toBeInstanceOf(PriorityQueue)
           expect(workerNode.tasksQueue.size).toBe(0)
           expect(workerNode.tasksQueue.maxSize).toBe(0)
+          expect(workerNode.tasksQueue.k).toBe(numberOfWorkers * 2)
         }
         await pool.destroy()
       },